Completed functional components for frontend & plugged-in with backend

+++ server/models/entry.go
@@ -7,7 +7,7 @@ import (
 type Entry struct {
 	ID          primitive.ObjectID `bson:"id"`
 	Dish        *string            `json:"dish"`
-	Fat         *float64           `json:"fat"`
+	Fat         *int64             `json:"fat"`
 	Ingredients *string            `json:"ingredients"`
 	Calories    *string            `json:"calories"`
 }
